Avoid using SerializedClientState in UI Move group-store and key-package-store out to "store" folder Move welcome and group methods out of "protocol.ts" file
