+++ examples/src/lib/key-package-store.ts
@@ -6,7 +6,7 @@ import {
   shareReplay,
   switchMap,
 } from "rxjs";
-import { KeyPackageStore } from "../../../src/core/key-package-store";
+import { KeyPackageStore } from "../../../src/store/key-package-store";
 import accounts from "./accounts";
 
 // Observable that triggers whenever the store changes
