+++ src/core/index.ts
@@ -1,13 +1,11 @@
 export * from "./capabilities.js";
-export * from "./client-state-storage.js";
-export * from "./client-state-utils.js";
+export * from "./client-state.js";
 export * from "./credential.js";
 export * from "./default-capabilities.js";
 export * from "./extensions.js";
-export * from "./group-store.js";
 export * from "./group.js";
 export * from "./key-package-relay-list.js";
-export * from "./key-package-store.js";
 export * from "./key-package.js";
 export * from "./marmot-group-data.js";
 export * from "./protocol.js";
+export * from "./welcome.js";
