+++ src/core/protocol.ts
@@ -1,11 +1,5 @@
+// Constants for the Marmot protocol
 import { defaultExtensionTypes } from "ts-mls";
-import { defaultClientConfig } from "ts-mls/clientConfig.js";
-import { marmotAuthService } from "./auth-service.js";
-import { bytesToHex } from "@noble/ciphers/utils.js";
-import { NostrEvent, UnsignedEvent } from "applesauce-core/helpers/event";
-import { finalizeEvent, generateSecretKey, getPublicKey } from "nostr-tools";
-import { encodeMlsMessage, type MLSMessage } from "ts-mls/message.js";
-import { encodeWelcome, type Welcome } from "ts-mls/welcome.js";
 
 /** The extension id for the last_resort key package extension for key packages */
 export const LAST_RESORT_KEY_PACKAGE_EXTENSION_TYPE = 0x000a;
@@ -60,12 +54,6 @@ export type ExtendedExtensionTypeName = keyof typeof extendedExtensionTypes;
 export type ExtendedExtensionTypeValue =
   (typeof extendedExtensionTypes)[ExtendedExtensionTypeName];
 
-/** Default ClientConfig for Marmot */
-export const defaultMarmotClientConfig = {
-  ...defaultClientConfig,
-  auth_service: marmotAuthService,
-};
-
 /**
  * Represents the decoded Marmot Group Data Extension structure.
  */
@@ -95,63 +83,3 @@ export const GROUP_EVENT_KIND = 445;
 
 /** Event kind for welcome events */
 export const WELCOME_EVENT_KIND = 444;
-
-/** Event kind for gift wrap events (NIP-59) */
-export const GIFT_WRAP_KIND = 1059;
-
-/**
- * Creates an unsigned Nostr event (kind 445) for a group commit message.
- *
- * @param commitMessage - The serialized MLS commit message
- * @param groupId - The 32-byte Nostr group ID (from MarmotGroupData)
- * @param pubkey - The sender's public key (hex string)
- * @param relays - Array of relay URLs for the group
- * @returns Unsigned Nostr event
- */
-export function createGroupEvent(
-  commitMessage: MLSMessage,
-  groupId: string,
-): NostrEvent {
-  const serializedMessage = encodeMlsMessage(commitMessage);
-  const content = bytesToHex(serializedMessage);
-  const secretKey = generateSecretKey();
-  const unsignedEvent = {
-    kind: GROUP_EVENT_KIND,
-    pubkey: getPublicKey(secretKey),
-    created_at: Math.floor(Date.now() / 1000),
-    content,
-    tags: [["h", groupId]],
-  };
-  return finalizeEvent(unsignedEvent, secretKey);
-}
-
-/**
- * Creates an unsigned Nostr event (kind 444) for a welcome message.
- *
- * @param welcomeMessage - The MLS welcome message
- * @param keyPackageId - The ID of the key package used for the add operation
- * @param pubkey - The sender's public key (hex string)
- * @param relays - Array of relay URLs for the group
- * @returns Unsigned Nostr event
- */
-export function createWelcomeEvent(
-  welcomeMessage: Welcome,
-  keyPackageId: string,
-  pubkey: string,
-  relays: string[],
-): UnsignedEvent {
-  // Serialize the welcome message according to RFC 9420
-  const serializedWelcome = encodeWelcome(welcomeMessage);
-  const content = bytesToHex(serializedWelcome);
-
-  return {
-    kind: WELCOME_EVENT_KIND,
-    pubkey,
-    created_at: Math.floor(Date.now() / 1000),
-    content,
-    tags: [
-      ["e", keyPackageId],
-      ["relays", ...relays],
-    ],
-  };
-}
