+++ src/index.ts
@@ -1,3 +1,3 @@
-// export all helpers
 export * from "./core/index.js";
+export * from "./store/index.js";
 export * from "./utils/index.js";
